Database Register Full Project

+++ src/main/java/hamitmizrak/dao/IDaoConnection.java
@@ -15,7 +15,7 @@ public interface IDaoConnection<T> {
     public ArrayList<T> list();
 
     // UPDATE
-    public T findById(int id);
+    public T findById(Long id);
 
     // UPDATE
     public void update(T t);
